requirements.txt
Flask==2.3.2
flask-cors==3.0.10
mediapipe==0.10.0
numpy==1.26.0
joblib==1.3.2
scikit-learn==1.3.2
Pillow==10.0.0
gunicorn==21.2.0
